fix: Chunk A code review findings — Rust + Python
Rust:
- adapter: path_contains disambiguator on ModuleMatcher::TypeAndName (GPT-2 c_proj)
- adapter: resolve() takes rank, generates probe_point format strings
- bridge: i64-before-f64 extraction in python_to_json_value
- dispatch: discover_execution_order + apply_execution_order in attach handler
- messages: model_handle + direction on HostStepRequest, model_handle on HostUpdateProbesRequest
- types: Default derive on StepDirection

Python:
- C1: tensor_to_bytes bf16 → fp16 conversion (NumPy has no bf16)
- C2: install_capture_hooks returns (handles, call_counts) for per-pass reset
- C3: compute_tensor_stats guards empty tensors
- C4: Mailbox.wait() accepts timeout, raises TimeoutError on expiry
- M5: sentinel hook returns None instead of output tensor
- I1: capture hook output type widened to Any
- I3: documented mailbox lock-state invariants (no code fix — protocol is correct)

+++ crates/rocket-surgeon-worker/src/bridge.rs
@@ -184,9 +184,9 @@ pub fn compute_tensor_stats(
 }
 
 fn python_to_json_value(obj: &Bound<'_, pyo3::PyAny>) -> anyhow::Result<serde_json::Value> {
-    if let Ok(v) = obj.extract::<f64>() {
+    if let Ok(v) = obj.extract::<i64>() {
         Ok(serde_json::Value::from(v))
-    } else if let Ok(v) = obj.extract::<i64>() {
+    } else if let Ok(v) = obj.extract::<f64>() {
         Ok(serde_json::Value::from(v))
     } else if let Ok(v) = obj.extract::<String>() {
         Ok(serde_json::Value::from(v))
